# swagger/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/contratos-generales": {
      "get": {
        "operationId": "ContratoGeneralController_findAll",
        "parameters": [
          {
            "name": "fields",
            "required": false,
            "in": "query",
            "description": "Campos a incluir en la respuesta (separados por comas)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Campo por el cual ordenar",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "orderBy",
            "required": false,
            "in": "query",
            "description": "Dirección del ordenamiento (ASC o DESC)",
            "schema": {
              "enum": [
                "ASC",
                "DESC"
              ],
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Número máximo de registros a retornar",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Número de registros a saltar",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "description": "Filtros en formato JSON. Puede incluir filtros para campos de relaciones usando notación punto. Ejemplo: {\"estados.estado_parametro_id\":1,\"solicitante.dependencia_solicitante_id\":101}",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "include",
            "required": false,
            "in": "query",
            "description": "Relaciones a incluir (separadas por comas). Ejemplo: estados,solicitante,lugarEjecucion",
            "schema": {
              "example": "estados,contratista",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de contratos generales",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContratoGeneral"
                  }
                }
              }
            }
          }
        },
        "summary": "Obtener todos los contratos generales",
        "tags": [
          "contratos-generales"
        ]
      },
      "post": {
        "operationId": "ContratoGeneralController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearContratoGeneralDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Contrato general creado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContratoGeneral"
                }
              }
            }
          }
        },
        "summary": "Crear un nuevo contrato general",
        "tags": [
          "contratos-generales"
        ]
      }
    },
    "/contratos-generales/{id}": {
      "get": {
        "operationId": "ContratoGeneralController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del contrato general",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "fields",
            "required": false,
            "in": "query",
            "description": "Campos a retornar (separados por comas)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Campo por el cual ordenar",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "orderBy",
            "required": false,
            "in": "query",
            "description": "Dirección del ordenamiento",
            "schema": {
              "enum": [
                "ASC",
                "DESC"
              ],
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Número de registros a retornar",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Número de registros a saltar",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "description": "Filtros dinámicos en formato JSON",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "include",
            "required": false,
            "in": "query",
            "description": "Relaciones a incluir (separadas por comas). Ejemplo: estados,contratista",
            "schema": {
              "example": "estados,contratista",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contrato general encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContratoGeneral"
                }
              }
            }
          },
          "404": {
            "description": "Contrato general no encontrado"
          }
        },
        "summary": "Obtener un contrato general por ID",
        "tags": [
          "contratos-generales"
        ]
      },
      "put": {
        "operationId": "ContratoGeneralController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del contrato general a actualizar",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActualizarContratoGeneralDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contrato general actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContratoGeneral"
                }
              }
            }
          },
          "404": {
            "description": "Contrato general no encontrado"
          }
        },
        "summary": "Actualizar un contrato general",
        "tags": [
          "contratos-generales"
        ]
      },
      "delete": {
        "operationId": "ContratoGeneralController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del contrato general a eliminar",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contrato general eliminado"
          },
          "404": {
            "description": "Contrato general no encontrado"
          }
        },
        "summary": "Eliminar un contrato general",
        "tags": [
          "contratos-generales"
        ]
      }
    },
...
    "/documentos-contratos/contrato/{contratoId}": {
      "get": {
        "operationId": "DocumentoContratoController_findByContratoId",
        "parameters": [
          {
            "name": "contratoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Devuelve los documentos contratos para el contrato especificado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentoContrato"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No se encontraron documentos contratos para el contrato especificado."
          }
        },
        "summary": "Obtener documentos contratos por id de contrato",
        "tags": [
          "documentos-contratos"
        ]
      }
    },
...
    "/contratistas/contrato/{contratoId}": {
      "get": {
        "operationId": "ContratistaController_findByContrato",
        "parameters": [
          {
            "name": "contratoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Devuelve el contratista para el contrato especificado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contratista"
                }
              }
            }
          },
          "404": {
            "description": "No se encontró contratista para el contrato especificado."
          }
        },
        "summary": "Obtener contratista por id de contrato",
        "tags": [
          "contratistas"
        ]
      }
    },
...
    "/estados-contrato/contrato/{contratoId}": {
      "get": {
        "operationId": "EstadoContratoController_findByContratoGeneral",
        "parameters": [
          {
            "name": "contratoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Devuelve los estados de contrato para el contrato especificado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EstadoContrato"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No se encontraron estados de contrato para el contrato especificado."
          }
        },
        "summary": "Obtener estados de contrato por id de contrato",
        "tags": [
          "estados-contrato"
        ]
      }
    },
...
    "/lugares-ejecucion/contrato/{contratoId}": {
      "get": {
        "operationId": "LugarEjecucionController_findByContratoGeneralId",
        "parameters": [
          {
            "name": "contratoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Devuelve los lugares de ejecución para el contrato especificado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LugarEjecucion"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No se encontraron lugares de ejecución para el contrato especificado."
          }
        },
        "summary": "Obtiene el lugar de ejecución por id de contrato",
        "tags": [
          "lugares-ejecucion"
        ]
      }
    },
...
    "/cdp/contrato/{contratoId}": {
      "get": {
        "operationId": "CdpController_findByContrato",
        "parameters": [
          {
            "name": "contratoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Devuelve los CDPs para el contrato especificado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Cdp"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No se encontraron CDPs para el contrato especificado."
          }
        },
        "summary": "Obtener CDPs por ID de contrato",
        "tags": [
          "cdp"
        ]
      }
    },
...
    "/solicitantes/contrato/{contratoId}": {
      "get": {
        "operationId": "SolicitanteController_findByContratoGeneralId",
        "parameters": [
          {
            "name": "contratoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Devuelve los lugares de ejecución para el contrato especificado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SolicitanteEntity"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No se encontraron lugares de ejecución para el contrato especificado."
          }
        },
        "summary": "Obtiene el lugar de ejecución por id de contrato",
        "tags": [
          "solicitantes"
        ]
      }
    }
  },
  "info": {
    "title": "Gestion Contractual CRUD",
    "description": "API para la gestión contractual",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "ContratoGeneral": {
        "type": "object",
        "properties": {}
      },
      "CrearContratoGeneralDto": {
        "type": "object",
        "properties": {
          "tipoCompromisoId": {
            "type": "number",
            "example": 1,
            "description": "Id Parámetros CRUD: Tipo Compromiso. 1. Convenio. 2. Contrato 3. Orden"
          },
          "tipoContratoId": {
            "type": "number",
            "example": 1,
            "description": "Id Parámetros CRUD: Tipo Contrato. 1. Orden de Servicio, 2. Orden de Compra"
          },
          "perfilContratistaId": {
            "type": "number",
            "example": 1,
            "description": "Id Parámetros CRUD: Perfil Contratista. 1. Asistencial, 2. Técnico, 3. Profesional, 4. Especializado, ..."
          },
          "aplicaPoliza": {
            "type": "boolean",
            "example": true,
            "description": "Aplica Poliza"
          },
          "modalidadSeleccionId": {
            "type": "number",
            "example": 1,
            "description": "Id Parámetros CRUD: Modalidad Selección. 1. Licitación Pública, 2. Selección Abreviada, ..."
          },
          "tipologiaEspecificaId": {
            "type": "number",
            "example": 30,
            "description": "Id Parámetros CRUD: Tipología Específica. 30. Servicios de Mantenimiento y Reparación, ..."
          },
          "regimenContratacionId": {
            "type": "number",
            "example": 1,
            "description": "Id Parámetros CRUD: Regimen Contratación. 1. Ley 80, 2. Régimen Privado, 3. Convenio Ley 489, ..."
          },
          "procedimientoId": {
            "type": "number",
            "example": 1,
            "description": "Id Parámetros CRUD: Procedimiento. 1. Subasta Inversa, 2. Mínima Cuantía, 3. Concurso de Méritos ..."
          },
          "plazoEjecucion": {
            "type": "number",
            "example": 30,
            "description": "Cantidad de días, meses o años para la ejecución del contrato"
          },
          "unidadEjecutoraId": {
            "type": "number",
            "example": 1,
            "description": "Id Parámetros CRUD: Unidad de Tiempo. 1. Días, 2. Meses, 3. Años"
          }
        },
        "required": [
          "tipoCompromisoId",
          "tipoContratoId",
          "perfilContratistaId",
          "aplicaPoliza",
          "modalidadSeleccionId",
          "tipologiaEspecificaId",
          "regimenContratacionId",
          "procedimientoId",
          "plazoEjecucion",
          "unidadEjecutoraId"
        ]
      },
      "ActualizarContratoGeneralDto": {
        "type": "object",
        "properties": {}
      },
...
      "DocumentoContrato": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "description": "Identificador único del documento"
          },
          "tipo_documento_id": {
            "type": "number",
            "example": 1,
            "description": "ID Tipo Documento - Parámetros CRUD"
          },
          "documento_enlace": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "description": "Enlace al documento"
          },
          "documento_id": {
            "type": "number",
            "example": 1,
            "description": "Documento ID - Gestor Documental"
          }
        },
        "required": [
          "id",
          "documento_id"
        ]
      },
...
      "Contratista": {
        "type": "object",
        "properties": {}
      },
...
      "EstadoContrato": {
        "type": "object",
        "properties": {}
      },
...
      "LugarEjecucion": {
        "type": "object",
        "properties": {}
      },
...
      "Cdp": {
        "type": "object",
        "properties": {}
      },
...
      "SolicitanteEntity": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}